+++ apps/openmw/mwworld/timestamp.cpp
@@ -33,7 +33,7 @@ namespace MWWorld
 
         mHour = static_cast<float> (std::fmod (hours, 24));
 
-        mDay += hours / 24;
+        mDay += static_cast<int>(hours / 24);
 
         return *this;
     }
