+++ components/terrain/defaultworld.cpp
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) 2015 scrawl <scrawl@baseoftrash.de>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 #include "defaultworld.hpp"
 
 #include <OgreAxisAlignedBox.h>
@@ -87,8 +108,8 @@ namespace Terrain
 
         storage->getBounds(mMinX, mMaxX, mMinY, mMaxY);
 
-        int origSizeX = mMaxX-mMinX;
-        int origSizeY = mMaxY-mMinY;
+        int origSizeX = static_cast<int>(mMaxX - mMinX);
+        int origSizeY = static_cast<int>(mMaxY - mMinY);
 
         // Dividing a quad tree only works well for powers of two, so round up to the nearest one
         int size = nextPowerOfTwo(std::max(origSizeX, origSizeY));
@@ -103,7 +124,7 @@ namespace Terrain
         LayersRequestData data;
         data.mPack = getShadersEnabled();
 
-        mRootNode = new QuadTreeNode(this, Root, size, Ogre::Vector2(centerX, centerY), NULL);
+        mRootNode = new QuadTreeNode(this, Root, static_cast<float>(size), Ogre::Vector2(centerX, centerY), NULL);
         buildQuadTree(mRootNode, data.mNodes);
         //loadingListener->indicateProgress();
         mRootNode->initAabb();
@@ -139,7 +160,7 @@ namespace Terrain
             float minZ,maxZ;
             Ogre::Vector2 center = node->getCenter();
             float cellWorldSize = getStorage()->getCellWorldSize();
-            if (mStorage->getMinMaxHeights(node->getSize(), center, minZ, maxZ))
+            if (mStorage->getMinMaxHeights(static_cast<float>(node->getSize()), center, minZ, maxZ))
             {
                 Ogre::AxisAlignedBox bounds(Ogre::Vector3(-halfSize*cellWorldSize, -halfSize*cellWorldSize, minZ),
                                     Ogre::Vector3(halfSize*cellWorldSize, halfSize*cellWorldSize, maxZ));
@@ -254,7 +275,7 @@ namespace Terrain
 
             LoadResponseData* responseData = new LoadResponseData();
 
-            getStorage()->fillVertexBuffers(node->getNativeLodLevel(), node->getSize(), node->getCenter(), getAlign(),
+            getStorage()->fillVertexBuffers(node->getNativeLodLevel(), static_cast<float>(node->getSize()), node->getCenter(), getAlign(),
                                             responseData->mPositions, responseData->mNormals, responseData->mColours);
 
             return OGRE_NEW Ogre::WorkQueue::Response(req, true, Ogre::Any(responseData));
