+++ components/terrain/quadtreenode.cpp
@@ -1,3 +1,24 @@
+/*
+ * Copyright (c) 2015 scrawl <scrawl@baseoftrash.de>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 #include "quadtreenode.hpp"
 
 #include <OgreSceneManager.h>
@@ -114,7 +135,7 @@ QuadTreeNode::QuadTreeNode(DefaultWorld* terrain, ChildDirection dir, float size
     : mMaterialGenerator(NULL)
     , mIsDummy(false)
     , mSize(size)
-    , mLodLevel(Log2(mSize))
+    , mLodLevel(Log2(static_cast<int>(mSize)))
     , mBounds(Ogre::AxisAlignedBox::BOX_NULL)
     , mWorldBounds(Ogre::AxisAlignedBox::BOX_NULL)
     , mDirection(dir)
@@ -415,7 +436,7 @@ void QuadTreeNode::updateIndexBuffers()
         // Fetch a suitable index buffer (which may be shared)
         size_t ourLod = getActualLodLevel();
 
-        int flags = 0;
+        unsigned int flags = 0;
 
         for (int i=0; i<4; ++i)
         {
@@ -436,7 +457,7 @@ void QuadTreeNode::updateIndexBuffers()
             if (lod > 0)
             {
                 assert (lod - ourLod < (1 << 4));
-                flags |= int(lod - ourLod) << (4*i);
+                flags |= static_cast<unsigned int>(lod - ourLod) << (4*i);
             }
         }
         flags |= 0 /*((int)mAdditionalLod)*/ << (4*4);
